+++ packages/bylaw_credo/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Update `Bylaw.Credo.Check.Testing.NoGlobalStateInTests` to recommend stable
+  environment-specific configuration.
 - Add `Bylaw.Credo.Check.Elixir.PreferBlockIf` to require block syntax for
   local and qualified Kernel `if` expressions.
 - Add `Bylaw.Credo.Check.Elixir.SimpleTaggedTupleValues` to require complex